+++ spec/lite/ruby/open_struct_spec.rb
@@ -3,6 +3,9 @@
 require 'spec_helper'
 
 RSpec.describe OpenStruct do
+  let(:h1) do
+    { table: { name: 'bob', age: 60, gender: :M } }
+  end
   let(:o1) do
     described_class.new(name: 'bob', age: 60) do |o|
       o.gender = :M
@@ -25,7 +28,7 @@ RSpec.describe OpenStruct do
 
   describe '#attributes' do
     it 'to be { ... }' do
-      expect(o1.attributes).to eq(name: 'bob', age: 60, gender: :M)
+      expect(o1.attributes).to eq(h1[:table])
     end
   end
 
@@ -34,6 +37,27 @@ RSpec.describe OpenStruct do
       o1.replace(name: 'tom', age: 28)
 
       expect(o1.name).to eq('tom')
+    end
+  end
+
+  describe '#to_hash' do
+    it 'to be { ... } with table key' do
+      expect(o1.to_hash).to eq(h1)
+    end
+
+    it 'to be { ... } without table key' do
+      expect(o1.to_hash(table: false)).to eq(h1[:table])
+    end
+  end
+
+
+  describe '#to_json' do
+    it 'to be { ... } with table key' do
+      expect(o1.to_json).to eq(h1.to_json)
+    end
+
+    it 'to be { ... } without table key' do
+      expect(o1.to_json(table: false)).to eq(h1[:table].to_json)
     end
   end